This is a small expansion of the normal little crab program.
You can change the amount of players in the world class.
There are images for 8 players but only controls for 4.
If you want to add more controls you can do it in the crab class.
There are some powerups that spawn randomly and give permanent effects to you or other players.
You can add new powerups if you want to just add the image and add it in the powerup calss,
add the effect it should have in the crab class and change the spawnpowerup method in the world.

Before you run the program again give it some time since it has to recompile the changes.

Signed-off-by: noa1510 <14131388+noah1510@users.noreply.github.com>